+++ src/pptx/__init__.py
@@ -21,6 +21,7 @@ from pptx.parts.slide import (
     SlideLayoutPart,
     SlideMasterPart,
     SlidePart,
+    ThemePart,
 )
 
 if TYPE_CHECKING:
@@ -45,6 +46,7 @@ content_type_to_part_class_map: dict[str, type[Part]] = {
     CT.PML_SLIDE: SlidePart,
     CT.PML_SLIDE_LAYOUT: SlideLayoutPart,
     CT.PML_SLIDE_MASTER: SlideMasterPart,
+    CT.OFC_THEME: ThemePart,
     CT.DML_CHART: ChartPart,
     CT.BMP: ImagePart,
     CT.GIF: ImagePart,
@@ -89,6 +91,7 @@ del (
     SlidePart,
     SlideLayoutPart,
     SlideMasterPart,
+    ThemePart,
     PresentationPart,
     CT,
     PartFactory,